Refactor performance test for full automation
The `test_speed.py` script was a semi-manual script with hardcoded values and placeholder data. This commit refactors it into a fully automated performance testing tool.

Key changes:
- Renamed `test_speed.py` to `performance_test.py` to align with the project's documentation.
- The script now automates the setup and teardown of the test environment, including starting the Hardhat node and deploying the contract.
- Hardcoded contract addresses and ABIs have been replaced with dynamic loading from build artifacts.
- Placeholder values for UI and frontend compilation tests, which cannot be reliably measured in this environment, have been removed.
- The script now measures and reports the Hardhat server startup time and contract deployment time.
