+++ processing.md
@@ -1,8 +1,6 @@
-Frontend_Compilation: Latency: 3.5000s Speed: 0.2857 m/s
-Create_Bounty: Latency: 0.0225s Speed: 44.3959 m/s
-Claim_Bounty: Latency: 0.0170s Speed: 58.8261 m/s
-Metamask_Approve: Latency: 0.0165s Speed: 60.5020 m/s
-Metamask_Cancel_Bounty: Latency: 0.0142s Speed: 70.6480 m/s
-View_Detail_Page: Latency: 0.8000s Speed: 1.2500 m/s
-Hardhat_Server_Startup: Latency: 5.0000s Speed: 0.2000 m/s
-Bounty_Card_Display: Latency: 0.5000s Speed: 2.0000 m/s
+Hardhat_Server_Startup: Latency: 5.5329s Speed: 0.1807 m/s
+Contract_Deployment: Latency: 3.4236s Speed: 0.2921 m/s
+Create_Bounty: Latency: 0.0417s Speed: 23.9622 m/s
+Claim_Bounty: Latency: 0.0260s Speed: 38.4492 m/s
+Metamask_Approve: Latency: 0.0277s Speed: 36.1497 m/s
+Metamask_Cancel_Bounty: Latency: 0.0241s Speed: 41.5376 m/s
