+++ src/artifacts/contracts/BonusEscrow.sol/BonusEscrow.dbg.json
@@ -1,4 +1,4 @@
 {
   "_format": "hh-sol-dbg-1",
-  "buildInfo": "..\\..\\build-info\\64ef1334a592d956d73f7abe39d3bdf2.json"
+  "buildInfo": "../../build-info/f2ac214ca5feb8f9427d9cbc0774b381.json"
 }
